+++ src/domain/journey/opportunity/pages/OpportunityCommunications/OpportunityCommunicationsPage.tsx
@@ -4,7 +4,7 @@ import CommunityUpdatesPage, {
 } from '@/domain/platform/admin/community/CommunityUpdatesPage';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
-import SubspaceSettingsLayout from '@/domain/platform/admin/subspace/SubspaceSettingsLayout';
+import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
 
 interface OpportunityCommunicationsPageProps extends SettingsPageProps, CommunityUpdatesPageProps {}
