+++ src/domain/journey/opportunity/pages/OpportunitySettings/OpportunitySettingsPage.tsx
@@ -1,7 +1,7 @@
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
-import SubspaceSettingsLayout from '@/domain/platform/admin/subspace/SubspaceSettingsLayout';
 import OpportunitySettingsView from './OpportunitySettingsView';
+import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
 
 const OpportunitySettingsPage = ({ routePrefix = '../' }: SettingsPageProps) => (
   <SubspaceSettingsLayout currentTab={SettingsSection.SpaceSettings} tabRoutePrefix={routePrefix}>
