+++ src/domain/journey/space/pages/SpaceSettings/SpaceSettingsPage.tsx
@@ -5,8 +5,8 @@ import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLa
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import useUrlResolver from '@/main/routing/urlResolver/useUrlResolver';
 import { Skeleton } from '@mui/material';
-import SubspaceSettingsLayout from '@/domain/platform/admin/subspace/SubspaceSettingsLayout';
 import { SpaceLevel } from '@/core/apollo/generated/graphql-schema';
+import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
 
 const SpaceSettingsPage: FC<SettingsPageProps> = ({ routePrefix = '../' }) => {
   const { spaceLevel, loading } = useUrlResolver();
