+++ src/domain/journey/subspace/pages/SubspaceSubspaces/SubspaceSubspacesPage.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
-import SubspaceSettingsLayout from '@/domain/platform/admin/subspace/SubspaceSettingsLayout';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import OpportunityList from '@/domain/journey/opportunity/pages/OpportunityList';
+import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
 
 const ChallengeOpportunitiesPage: FC<SettingsPageProps> = ({ routePrefix }) => {
   return (
