--- src/domain/platform/admin/space/SpaceMembershipPreferenceQuery.graphql
@@ -1,1 +1,0 @@
-# this should contain space membership preferences when we are able to filter on the field
