+++ src/domain/space/routing/toReview2/nonSpaceAdminRedirect/SpacePrivileges.graphql
