+++ src/domain/space/routing/toReviewAdmin/ChallengesRoute.tsx
@@ -4,7 +4,7 @@ import { Error404 } from '@/core/pages/Errors/Error404';
 import { nameOfUrl } from '@/main/routing/urlParams';
 import SubspaceProvider from '@/domain/journey/subspace/context/SubspaceProvider';
 import SubspaceListPage from '@/domain/space/layout/tabbedLayout/Tabs/SpaceSubspaces/SubspaceListPage';
-import { ChallengeRoute } from '@/domain/journey/settings/routes/ChallengeRoute';
+import { ChallengeRoute } from '@/domain/space/routing/toReview2/ChallengeRoute';
 
 const ChallengesRoute = () => {
   return (
